+++ src/mpack/mpack-platform.h
@@ -748,6 +748,7 @@ size_t mpack_strlen(const char* s);
 
 /* Debug logging */
 #if 0
+    #include <stdio.h>
     #define mpack_log(...) (printf(__VA_ARGS__), fflush(stdout))
 #else
     #define mpack_log(...) ((void)0)